feat: add support for SSL connections to database
Setting the new environment variable `MYSQL_SSL` to `True` will ensure that any
connection to the database server is secured. The path to the CA certificates is
specific for Alpine (as that is used in the Dockerfile).

For local deployments or development, setting the environment variable to
`False` allows normal unsecured connections.